# gateway/openapi.yaml
openapi: "3.0.3"
info:
  title: MoCKA Gateway API
  description: |
    MoCKA Multi-AI Gateway — Copilot Studio Custom Connector 登録用OpenAPI定義。
    全AIがMoCKAの知識基盤にアクセス・記録できるREST APIを提供する。
  version: "1.1"
  contact:
    name: MoCKA Project
    email: nsjpkimura@gmail.com
servers:
  - url: https://gateway.nsjp.org
    description: Production (Cloudflare Tunnel, TODO_266)
security:
  - ApiKeyAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-MoCKA-Key
  schemas:
    Context:
      type: object
      properties:
        phase:
          type: string
          description: MoCKAの現在フェーズ
        goal:
          type: string
          description: 直近の最優先目標
        last_decision:
          type: string
          description: 最新の意思決定内容
        active_todo:
          type: array
          items:
            $ref: "#/components/schemas/Todo"
        essence_summary:
          type: string
          description: MoCKA本質サマリー（compact除く）
        recent_events:
          type: array
          items:
            $ref: "#/components/schemas/Event"
        meta:
          $ref: "#/components/schemas/Meta"
    Todo:
      type: object
      properties:
        id:       { type: string }
        title:    { type: string }
        priority: { type: string, enum: [最高, 高, 中, 低] }
        status:   { type: string }
    Event:
      type: object
      properties:
        id:            { type: string }
        title:         { type: string }
        short_summary: { type: string }
        when:          { type: string, format: date-time }
        what_type:     { type: string }
    Meta:
      type: object
      properties:
        version:      { type: string }
        mode:         { type: string, enum: [compact, standard, extended] }
        generated_at: { type: string, format: date-time }
        size_bytes:   { type: integer }
    Actor:
      type: object
      required: [vendor, source]
      properties:
        vendor:  { type: string, example: OpenAI }
        model:   { type: string, example: GPT-5.5 }
        version: { type: string }
        runtime: { type: string, example: ChatGPT }
        source:  { type: string, example: Orchestra }
    EventRequest:
      type: object
      required: [title, description, actor, timestamp, nonce, request_id]
      properties:
        title:       { type: string, maxLength: 100 }
        description: { type: string }
        tags:
          type: array
          items: { type: string }
        actor:
          $ref: "#/components/schemas/Actor"
        request_id:
          type: string
          format: uuid
          description: UUIDv4 — 重複防止
        timestamp:
          type: string
          format: date-time
          description: ISO8601 UTC — ±5分以内のみ受付
        nonce:
          type: string
          description: 16バイトランダムHex — 10分間有効
        hmac_sig:
          type: string
          description: HMAC-SHA256署名（任意）
...
paths:
...
  /api/v1/context:
    get:
      operationId: getContext
      summary: MoCKAコンテキスト取得
      description: |
        AIプロンプトに注入するMoCKAの現在状態を返す。
        mode パラメータで情報量を制御できる。
      parameters:
        - name: mode
          in: query
          required: false
          schema:
            type: string
            enum: [compact, standard, extended]
            default: standard
          description: compact（〜2KB）/ standard（〜8KB）/ extended（〜50KB）
      responses:
        "200":
          description: コンテキスト情報
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Context" }
        "401":
          description: APIキー未指定
...
  /api/v1/event:
    post:
      operationId: postEvent
      summary: イベント記録
      description: |
        MoCKAにイベントを書き込む。
        actor構造体必須。request_id + nonce + timestampでReplay Attack防止。
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EventRequest" }
      responses:
        "201":
          description: 記録成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:   { type: string }
                  event_id: { type: string }
        "400":
          description: バリデーションエラー
        "401":
          description: 認証失敗（APIキー / timestamp超過）
        "409":
          description: nonce重複（Replay Attack検出）